Move flags and parsing to the same place
